requirements.txt
openai
streamlit
google-cloud-bigquery
google-cloud-bigquery-storage
pandas-gbq